add dpdk_enabled parameter
the difference of ovs upgrade between dpdk-compute
and controller nodes is whether to configure dpdk
port. This patch makes the dpdk task configurable
so that it can be used on both compute and controller
nodes. Default dpdk_enabled to false.

There is case that some compute nodes are not using
dpdk ports, these nodes can be treated the same as
controller nodes by setting dpdk_enabled to false.